+++ src/income.spec.ts
@@ -1,7 +1,6 @@
 import Event from "./events";
 import Player from "./player";
 import { expect } from "chai";
-import { combinations } from "./income";
 import { Power } from "./player-data";
 import { BrainstoneArea } from "./enums";
 
@@ -93,7 +92,7 @@ describe("IncomeSelection", () => {
         player.data.brainstone = test.brainstone;
         player.loadEvents(test.events);
 
-        expect(player.getIncomeSelection().remainingChargesAfterIncome).to.equal(test.expected);
+        expect(player.incomeSelection().remainingChargesAfterIncome).to.equal(test.expected);
       });
     }
   });
@@ -132,7 +131,7 @@ describe("IncomeSelection", () => {
         player.settings.autoIncome = test.give.autoIncome;
         player.loadEvents(test.give.events);
 
-        const incomeSelection = player.getIncomeSelection();
+        const incomeSelection = player.incomeSelection();
         expect(incomeSelection.needsManualSelection).to.equal(test.want.needsManualSelection);
         expect(incomeSelection.autoplayEvents()).to.deep.equal(test.want.events);
       });
@@ -174,33 +173,7 @@ describe("IncomeSelection", () => {
         player.data.power = test.give.power;
         player.loadEvents(test.give.events);
 
-        expect(player.getIncomeSelection().autoplayEvents()).to.deep.equal(test.want.events);
-      });
-    }
-  });
-
-  describe("combinations", () => {
-    const tests: {
-      name: string;
-      give: number[];
-      want: number[][];
-    }[] = [
-      {
-        name: "empty",
-        give: [],
-        want: [[]],
-      },
-      {
-        name: "1,2",
-        give: [1, 2],
-        want: [[], [1], [2], [2, 1]],
-      },
-    ];
-
-    for (const test of tests) {
-      it(test.name, () => {
-        const target = combinations(test.give);
-        expect(target).to.deep.equal(test.want);
+        expect(player.incomeSelection().autoplayEvents()).to.deep.equal(test.want.events);
       });
     }
   });
